chore: Move generateTicketNumbers and generateRandomNumber into useTicketsReducer
